26-11-2025/pyproject.toml
[project]
name = "26-11-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "inngest>=0.5.13",
    "llama-index-core>=0.14.8",
    "llama-index-readers-file>=0.5.5",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
